[fx] allow custom behavior for args, kwargs, and bool
This change makes it possible to subclass the tracer to add additional
behavior when you know something about the shape of the Proxy objects,
by overriding the defaults for how the tracer tries to make something iterable,
looks for keys for **kwargs, or tries to convert to a boolean.

An example test shows how this can be used to tag inputs with shapes.
It can also be used combined with create_node to do type propagation during
tracing to fullfil requests like iter.

ghstack-source-id: 38e2186fbc4a4ca660b81185aed6eefe8a58da0b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45193grammatically update index.rst (#45801)